requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
Flask==2.3.3
Flask-RESTful==0.3.10
